Tidy up party to only post case event once when enrolling and no active enrolments
