+++ ras_party/controllers/account_controller.py
@@ -8,6 +8,7 @@ from requests import HTTPError
 from werkzeug.exceptions import BadRequest, Conflict, InternalServerError,  NotFound
 
 from ras_party.clients.oauth_client import OauthClient
+from ras_party.controllers.case_controller import get_cases_for_casegroup, post_case_event
 from ras_party.controllers.iac_controller import disable_iac, request_iac
 from ras_party.controllers.notify_gateway import NotifyGateway
 from ras_party.controllers.queries import count_enrolment_by_survey_business
@@ -152,13 +153,10 @@ def change_respondent_enrolment_status(payload, session):
     session.commit()
 
     # If no enrolments are remaining for business/survey
-    # then send NO_ACTIVE_ENROLMENTS case event for each relevant B case
-    casegroup_ids = get_business_survey_casegroups(survey_id, business_id)
+    # then send NO_ACTIVE_ENROLMENTS case event
     enrolment_count = count_enrolment_by_survey_business(business_id, survey_id, session)
     if not enrolment_count:
-        logger.info('No active enrolments', business_id=business_id, survey_id=survey_id)
-        for case in get_cases_for_casegroups(casegroup_ids, business_id):
-            post_case_event(case['id'],
+            post_case_event(get_case_id_for_business_survey(survey_id, business_id),
                             category='NO_ACTIVE_ENROLMENTS',
                             desc='No active enrolments remaining for case')
 
@@ -583,9 +581,7 @@ def add_new_survey_for_respondent(payload, tran, session):
     disable_iac(enrolment_code, case_id)
 
     if count_enrolment_by_survey_business(survey_id, business_id, session) == 0:
-        casegroup_ids = get_business_survey_casegroups(survey_id, business_id)
-        for case in get_cases_for_casegroups(casegroup_ids, business_id):
-            post_case_event(case['id'], "RESPONDENT_ENROLED", "Respondent enroled")
+        post_case_event(case_id, "RESPONDENT_ENROLED", "Respondent enroled")
 
     # This ensures the log message is only written once the DB transaction is committed
     tran.on_success(lambda: logger.info('Respondent has enroled to survey for business',
@@ -644,9 +640,7 @@ def enrol_respondent_for_survey(respondent, session):
     logger.info('Pending enrolment for case_id', case_id=case_id)
     if count_enrolment_by_survey_business(survey_id=enrolment.survey_id, business_id=enrolment.business_id,
                                           session=session) == 0:
-        casegroup_ids = get_business_survey_casegroups(pending_enrolment.survey_id, pending_enrolment.business_id)
-        for case in get_cases_for_casegroups(casegroup_ids, pending_enrolment.business_id):
-            post_case_event(case['id'], "RESPONDENT_ENROLED", "Respondent enrolled")
+        post_case_event(case_id, "RESPONDENT_ENROLED", "Respondent enrolled")
     session.delete(pending_enrolment)
 
 
@@ -698,31 +692,6 @@ def request_survey(survey_id):
     return response.json()
 
 
-def post_case_event(case_id, category='Default category message', desc='Default description message'):
-    logger.debug('Posting case event', case_id=case_id)
-    case_svc = current_app.config['RAS_CASE_SERVICE']
-    case_url = f'{case_svc}/cases/{case_id}/events'
-    payload = {
-        'description': desc,
-        'category': category,
-        'createdBy': 'Party Service'
-    }
-
-    response = Requests.post(case_url, json=payload)
-    response.raise_for_status()
-    logger.debug('Successfully posted case event')
-    return response.json()
-
-
-def request_cases_for_respondent(respondent_id):
-    logger.debug('Retrieving cases for respondent', respondent_id=respondent_id)
-    url = f'{current_app.config["RAS_CASE_SERVICE"]}/cases/partyid/{respondent_id}'
-    response = Requests.get(url)
-    response.raise_for_status()
-    logger.debug('Successfully retrieved cases for respondent', respondent_id=respondent_id)
-    return response.json()
-
-
 def request_casegroups_for_business(business_id):
     logger.debug('Retrieving casegroups for business', business_id=business_id)
     url = f'{current_app.config["RAS_CASE_SERVICE"]}/casegroups/partyid/{business_id}'
@@ -756,16 +725,14 @@ def get_business_survey_casegroups(survey_id, business_id):
     return ce_casegroup_ids
 
 
-def get_cases_for_casegroups(casegroup_ids, case_party_id):
-    logger.debug('Retrieving cases for casegroups', case_party_id=case_party_id)
+def get_case_id_for_business_survey(survey_id, business_id):
+    logger.debug('Retrieving case for survey and business', survey_id=survey_id, business_id=business_id)
 
-    cases = request_cases_for_respondent(case_party_id)
-    # Filtering cases by survey/business
-    matching_cases = [case for case in cases
-                      if case['caseGroup']['id'] in casegroup_ids]
-    logger.debug('Successfully retrieved cases for casegroups',
-                 case_party_id=case_party_id)
-    return matching_cases
+    case_group_ids = get_business_survey_casegroups(survey_id, business_id)
+
+    cases = get_cases_for_casegroup(case_group_ids[0])
+
+    return cases[0]['id']
 
 
 def _is_valid(payload, attribute):
